fix(checker): statement walk cedes overload ambient-agreement to the symbol pass (#16742)
The adjacency-based TS2384 arms in the statement walk double-reported
against the new symbol-level flag-agreement pass and could not yield
export precedence (tsc reports only TS2383 when a declaration deviates
on both axes) nor blame every signature of a run. The walk keeps its
control flow (ambient signatures still skip TS2391 grouping) and the
symbol pass owns the diagnostic. The all-function TS2652 test pinned
the pre-oracle expectation; typescript@7.0.2 reports TS2383+TS2391 for
that shape and no merge-family diagnostic.
